+++ Assets/Scripts/UI/SliderTimer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6e54473f6afbe3e4aac9e773245c48df
+guid: be754b3e9ca64094bad43f25bdc90595
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
